cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy yf225 chenyang78 kadeng muchulee8 ColinPeppler amjames desertfire chauhang aakhundov

Allow mutations mutations for subclasses that are non-contiguous.

Changes:

Removing assert in collect_metadata_analysis

Main requested testcase:
Compilation of NJT.index_put()

Adding test in test_nestedtensor.py, that compiles NJT.index_put()

It is  decomposed to NJT split,unbind, which  needed additional `torch._check`, `torch._check_is_size` for NJT.unbind()  and guard_size_oblivious() usage in _meta_registrations and _inductor/lowering.py.


Special case:
If tangent is mutated outside of the graph, it does not participate in backward graph. Autograd in this case will set this tangent to zeros tensor.

We handle it separately in CompiledFunction.backward: not doing any processing for this tangent and broadcast to number of expected subclass unwrapped arguments.

disabling for dynamo 2 tests:
1/ For nested tensor - symbolic shapes issue on nested_tensor index operation that does splits [0, 0, 0] - there is a failure with "pending unbacked symints". This PR does not add more .tolist()/item() ops than it was before.

2/ As we do not fail with exception in collect_metadata_analysis new paths for dynamo started working and it started failing with smth strange that set_ in storage_offset (because of test for views) handling updates storage "cpu" -> "meta"




[ghstack-poisoned]

+++ test/cpp_extensions/open_registration_extension/pytorch_openreg/csrc/OpenRegHooks.cpp
@@ -186,7 +186,7 @@ struct OpenRegGuardImpl final : public c10::impl::DeviceGuardImplInterface {
   void destroyEvent(void* event, const c10::DeviceIndex device_index)
       const noexcept override {
     py::gil_scoped_acquire acquire;
-    get_method("destroyEvent")(event, device_index);
+    get_method("destroyEvent")((int64_t)event, device_index);
   }
 
   /**
@@ -201,7 +201,7 @@ struct OpenRegGuardImpl final : public c10::impl::DeviceGuardImplInterface {
       const c10::DeviceIndex device_index,
       const c10::EventFlag flag) const override {
     py::gil_scoped_acquire acquire;
-    get_method("record")(event, stream, device_index, flag);
+    get_method("record")((int64_t)event, stream, device_index, (int64_t)flag);
   }
 
   /**
@@ -214,7 +214,7 @@ struct OpenRegGuardImpl final : public c10::impl::DeviceGuardImplInterface {
    */
   void block(void* event, const c10::Stream& stream) const override {
     py::gil_scoped_acquire acquire;
-    get_method("block")(event, stream);
+    get_method("block")((int64_t)event, stream);
   }
 
   /**
@@ -225,7 +225,7 @@ struct OpenRegGuardImpl final : public c10::impl::DeviceGuardImplInterface {
    */
   bool queryEvent(void* event) const override {
     py::gil_scoped_acquire acquire;
-    return get_method("queryEvent")(event).cast<bool>();
+    return get_method("queryEvent")((int64_t)event).cast<bool>();
   }
 
   /**
@@ -261,7 +261,7 @@ struct OpenRegGuardImpl final : public c10::impl::DeviceGuardImplInterface {
    */
   void synchronizeEvent(void* event) const override {
     py::gil_scoped_acquire acquire;
-    get_method("synchronizeEvent")(event);
+    get_method("synchronizeEvent")((int64_t)event);
   }
 
   /**
@@ -281,7 +281,7 @@ struct OpenRegGuardImpl final : public c10::impl::DeviceGuardImplInterface {
   double elapsedTime(void* event1, void* event2, const c10::DeviceIndex device_index)
       const override {
     py::gil_scoped_acquire acquire;
-    return get_method("elapsedTime")(event1, event2, device_index).cast<double>();
+    return get_method("elapsedTime")((int64_t)event1, (int64_t)event2, device_index).cast<double>();
   }
 };
